Move spindle save widget to main settings widget file
